Minimalist Clustered Boxplot in R
Create a minimalist boxplot that displays medians of a dataset containing three measured y variables across a factor with x conditions. Designed with Edward Tufte's advice in mind.

The reshape library is used for data preparation to melt the data into long form, and then ggplot2 is used to create the plot.